Update restore to delete current connections to database
Postgres require that the database doesn't have connections to be deleted. Otherwise it will raise an error. On postgres 13+ the dropdb command have the --force option that terminates the connections forcefully. For older versions of postgres we need to do that ourselves. I've updated the restore script to terminate the current connections to the ghostwriter database before deleting it.