Add user-defined fields to projects
This patch allows administrators to define additional "extra fields"
via the admin API to include additional deployment-specific data with
ghostwriter models. The data is available via the UI as well as in
report templates.

Define the extra fields by going to the Admin Panel, then the "Extra
Field Configurations" page, then selecting the relevant model. After
adding custom fields, they will be visible in detail and edit pages,
and exposed to the template engine via
`{{ <instance>.extra_fields.<internal_name> }}`.

This patch only adds extra fields support to the Project model, as a
starting point, but the relevant code is extensible to other models.

Internally, this is implemented as two new models: `ExtraFieldModel`,
which corresponds to each Django model and is installed via a fixture,
and `ExtraFieldSpec`, which defines and configures the fields. The
actual data for the extra fields is stored in a `JSONField` on the
relevant model. The patch includes form fields, form widgets, and
django-rest-api serializers that query the `ExtraFieldSpec` and
produces sub fields and widgets based on the declarations.