Added JS to replace evidence placeholders with previews
The Js tracks all report evidence in an `evidenceFiles` list variable with the placeholder and a value that is either HTML or a URL to get dynamic HTML. On page load, the JS loops over all extra field preview modals for rich text fields to find and replace evidence placeholders. The function pulls dynamic HTML from the `evidence_preview` route.