Change how extra field defaults work
Instead of being used instead of an empty value, the extra field default
is now the initial value on new objects. This allows users to use the
default value as a base to edit. It also preserves old reports, since
changing the default value will not change any existing values.