Fix SuspiciousFileOperation when archiving projects
GW was opening the file in the archive export location and passing it to
the FileField, which then reads the file path off of it and tries to use
it as the file. This throws an error if the archive location is absolute
or has a ".." component.

Instead, pass it in the zip buffer so that Django creates an accessible
file the normal way.