Clear report template if it no longer matches the template
This patch will clear the global default template or a report's default
template if that template's type or client is updated in a way that no
longer matches how its being used.

This also adds warning messages to the template edit form to note that
the template defaults will be cleared, and if the template is a global
default.