Oplog full-text-search take two
This patch enables searching of tags and extra fields. The downside
is that we can't build an FTS index, since the tags are in a different
table and the extra fields are user-editable.

In the future, we can look into storing the compiled SearchVector as a
field in the OplogEntry, updating it whenever the entry or its tags
change, and indexing that.