Properly escape form-related toast messages
Currently, the text for the toasts is provided in a script tag via
standard `{{ text }}` filter, which does not escape for javascript
properly and causes text with angle brackets in it to show the HTML
escapes in the toast, which is ugly.

This patch passes the text through the `escapejs` filter, which does
proper javascript escaping.