Add support for looping lists and tables in rich text jinja
Similar to docx templates, you can now use `{%li for ...}`,
`{%tr for ...}`, `{%p for ...}` and other variations to repeat sections
for lists, tables, and paragraphs. This works by finding the ancestor of
the jinja tag that matches the prefix, and replacing it and its contents
with the jinja tag without the prefix.