Make most optional text columns non-nullable
Allowing null and blank values in a text column generally isn't
recommended, since it allows two "blank" values with different
semantics. The Ghostwriter UI shouldn't fill anything with null, but
the graphql api may.

There are a few fields that allowed null but not blank values, making
them effectively required. This patch makes them non-nullable, and
replaces any nulls with placeholders.

This also simplifies client contact validation. The custom validation
would require the contact title and email only if a name was specified,
but since the name was a required field, that was always the case. This
patch removes that custom value and just makes those fields required.