Changed `get_reports` tag to `get_assignment_data` to reduce queries
The `get_reports` tag for the "My Active Reports" sidebar dropdown made a list of reports based on a user's project assignments. We repeated this for the project assignments under the "My Active Projects" dropdown. For projects, we referenced `if request.user.projectassignment_set.all`.

If a project was marked complete, we weren't showing it in the sidebar, but if a user's projects were all marked as complete the dropdown would appear to be broken (no "no assignments" message). Rather than repeat what we did for `get_reports` and continue making the same query twice, I condensed it all into one tag that returns a unique list of projects and reports.