Don't show client projects for users just assigned to a project
This is a breaking change in how permissions work. Currently, if an
operator is assigned to any project for a client, they get read access
to all of the client's other projects as well. This patch changes that
to only allow access to the project they've been assigned.

Management and users with client invitations still have full view
access.