Added error handling to catch potential `csv.Error`
Python `csv` limits cells to 128KB. A user might upload a file with a cell containing text larger than 128KB and trigger a `csv.Error`. This causes an unhandled 500 error. This size limit is the only known error scenario so this adds a toast message explaining the provided file may have one or more cells that are too large.

We can increase the max size with `csv.field_size_limit(sys.maxsize)`, but this limit seemed too high. It caused browser tab's to crash (the server side had no issues). We might find a middle ground, but 128KB has been big enough so far. If this becomes a common thing we can explore raising it.