Changed setting Content-Disposition to use new custom func for encoding
Django's generic `HttpResponse` does not handle RFC 5987 like the `FileResponse` subclass does. We use `HttpResponse` for report downloads and exports. If the filename contained a non-ascii character (e.g., Č in a client name), the response would send a bad `Content-Disposition` value and result in a file named `download.ext` or a random string of characters set by the browser. This patch adds `add_content_disposition_header()` to format a proper `Content-Disposition` header for non-ascii characters.