Add support for CVSS V4
Add separate fields for CVSS V4 - place who want to use CVSS V3 can
continue to do so and ignore the V4 fields.

Add a CVSS V4 calculator, from Red Hat's code, but refactored to use
local variables. The calculators are in separate dialogs, so users
may hide whichever one they are not using.

Places where a score is used (ex sorting or the finding list) will use
either score, with a preference for v4 if both are specified.