Upgrade to Django 4.2
Upgrade several dependendencies as well to support 4.2. Upgrade the
Postgres image version to latest stable. Remove the deprecated pytz and
move to zoneinfo.

Tests pass but I've only lightly used it. There may be other issues.

This requires a manual Postgres database upgrade. A PR for
ghostwriter-cli is available that helps automate the upgrade.