Fix oplog search breaking with very long entries
Postgres aborts queries that try to make a `tsvector` with too many
unique tokens. This patch reworks the FTS oplog query to limit each
individual field to avoid collecting too many tokens and running into
the error.

A drawback is that text after this cutoff point will not be searchable.

Fixes #557