Collaborative editing, take two
Instead of transitioning everything to YJS, which would result in a lot
of code needing to be migrated, this uses a sidecar Hocuspocus server
to handle collaborative editing, converting to/from YJS as needed.

The Hocuspocus communicates primairly via graphql, but this adds another
REST endpoint to authenticate users.

Production docker has not yet been updated with the changes.