Stateful model handelers in collab server
Instead of handlers being a simple load+save function, rework them to
instantiate a new instance of a handler subclass on document load. This
will be used for adding additional state to a handler, like an Apollo
subscription.

Hocuspocus does not have a "document context", so keep track of a map
of document names to handler instances ourselves.