Revert back to load/save collab handlers, reduce boilerplate
Undoes 5e231f6f595c47e473a06389e5e64421c36a8bc0. I was planning on
having the collab server watch for changes in dropdown options,
requiring persistence. However it was ultimately simpler to have the
frontend contact the GraphQL API itself, rendering the persistence
unneeded.

This also adds `simpleModelHandler`, which takes care of the boilerplate
of running the query and setting up the document, so that handlers
only need to provide the queries and functions for setting/extracting
things on the doc.