Fix FindingCreateViewTests failing
The error is pretty unhelpful, especially since it doesn't occur when
running the test by itself.

The create form seems to default to finding type 1 and severity 1, but
they didn't exist. Seems like checking FK constraints is delayed until
the end of the test case, and might be skipped if there's no follow up
test case to run.