Add table caption support to collab editor
Actually supporting `<caption>` inside of the table is not easy to
support - the table plugin assumes that each of its children is a row
in many places.

This wraps tables in a table-with-caption node that contains both.

It also employs some hack-ish DOM manipulation to convert TinyMCE tables
that use a <caption> element into this new format.