Fix collab editor JSON editing
The frontend and collab server were not on the same page on how the JSON
was stored in the yjs doc. With this patch, the JSON is now stored in
string form in the yjs doc, which is parsed when saving the doc to the
database.

Also allows passing data in the collab server from the load function
to the save function - at present its used for the extra field spec.