GQL endpoint for getting objects by tags
Includes `*_by_tag` for several models that get all models that include
the tag passed in as the query's parameter (and also that the querying
user has access to).

Adds a `user_viewable` method alongside the `user_can_*` to queryable
models that returns a queryset of the instances that the user is allowed
to view.