Extend collab editor JWT expiry time
Currently expires based on the `JWT_EXPIRATION_DELTA` setting which
defaults to 15 minutes, which isn't a lot of time to spend editing an
option. After this timeout, the GraphQL requests that the page makes
(most notably for evidences) won't work anymore.

This patch sets the token to expire after 24 hours, which should be
plenty of time for a session.