Add Project Collaborative Note
This collab-enabled free-form rich text field is meant for people
participating in a project to take whatever notes they need to. It is
not intended to be included in reports, though it is exposed to Jinja
nevertheless.

Includes some refactoring of the collab update view/template so that
collab editors can be included in templates that don't use the base
template.