Remove ability to reorder findings
Due to the collab server limitations, we can't change the severity from
outside the collab server without race conditions, so that part was
disabled, but findings were still orderable within a severity. But this
was confusing - people expected that dragging a finding would change
its severity when it wasn't supported.

So for now, just remove it entirely. We may want to revert this change
once we have a system to notify the collab editing server with django
side updates.