Remove finding evidence, move to report evidence
Evidences were initially associated with specific findings. Recently,
evidences gained the ability to be associated with an entire report
instead. This patch removes the ability to associate evidenes with
findings, allowing only report-associated evidences.

This should be more intuitive to use.

The migration will move existing finding-associated evidences to the
report of the findings. For most cases, this should not be an issue,
however, __this may cause problems if two evidences were created with
the same friendly name in the same report__.